text flow for audio & Sui matching
The last change in the text here actually censors the wording for Mion's
family businesses to "financial sector, real estate sector, apparel
industry, amusement industry" - which are kind of funny to look at as
polite ways of saying the pre-censored industries.

Leaving this one uncensored in the translation for now! (any opinions?)