Sui voiced text matching
regarding the last change to 10_2, the original didn't seem to translate
the term 補導 correctly, and indeed it was hard to get a sense of what was
meant from just dictionary definitions.
I found the following wikipedia page most informative
https://ja.wikipedia.org/wiki/%E8%A3%9C%E5%B0%8E
seems to be the class of law that is applied to juvenile crime and
includes powers of arrest, detention and interrogation. From the
context, I think it's fairly clear that Shion means that Mion was
detained more than once.