Update wata_010.txt to fix chronological translation error
In the context of the scene where I've changed a line, Keiichi is asked about what he did the night of the Watanagashi festival. In his thoughts, he says "We thought nobody knew... about what we did last night, but..."
The issue is that the Watanagashi festival was not last night, but the night prior; day 8, not day 10. I've changed the word "last" to "that".
"We thought nobody knew... about what we did that night, but..."